+++ web/src/stream_data.js
@@ -426,6 +426,10 @@ export function update_can_remove_subscribers_group_id(sub, can_remove_subscribe
     sub.can_remove_subscribers_group_id = can_remove_subscribers_group_id;
 }
 
+export function update_push_notifications_enabled(sub, push_notifications_enabled) {
+    sub.push_notifications_enabled = push_notifications_enabled;
+}
+
 export function receives_notifications(stream_id, notification_name) {
     const sub = sub_store.get(stream_id);
     if (sub === undefined) {