+++ web/tests/lib/events.js
@@ -51,6 +51,7 @@ exports.test_streams = {
         message_retention_days: null,
         stream_post_policy: 1,
         can_remove_subscribers_group_id: 2,
+        push_notifications_enabled: false,
     },
     test: {
         name: "test",
@@ -66,6 +67,7 @@ exports.test_streams = {
         message_retention_days: null,
         stream_post_policy: 1,
         can_remove_subscribers_group_id: 2,
+        push_notifications_enabled: false,
     },
 };